requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.7
django-allauth==0.51.0
django-bootstrap-v5==1.0.11
django-bootstrap5==22.1
django-import-export==3.0.2
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
h11==0.13.0
httplib2==0.21.0
idna==3.4
lxml==4.9.1
MarkupPy==1.14
numpy==1.23.4
oauth2client==4.1.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.1
Pillow==9.3.0
protobuf==4.21.12
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.3.0
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
